refactor: context 폴더 추가, 파일 이름변경: appwrapper-> contextprovider
