+++ src/App.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import { useContext } from "react";
 import Header from "./component/header/Header";
-import UserContext from "./userContext";
+import UserContext from "./context/userContext";
 import Ranking from "./component/atom/ranking";
 import GameBoard from "./component/atom/GameBoard";
