+++ src/component/atom/ranking.tsx
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import UserContext from "../../userContext";
+import UserContext from "../../context/userContext";
 import { useContext } from "react";
 
 const RankTable = styled.table`
