+++ src/component/header/Header.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "@emotion/styled";
 import LeftMenu from "./LeftMenu";
 import RightMenu from "./RightMenu";
-import UserContext from "../../userContext";
+import UserContext from "../../context/userContext";
 import { useContext } from "react";
 
 const StyledHeader = styled.header`
