+++ src/main.tsx
@@ -1,10 +1,10 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
-import AppWrapper from "./AppWrapper.tsx";
+import ContextProvider from "./context/contextProvider";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <AppWrapper />
+    <ContextProvider />
   </StrictMode>
 );
